Refactor property setup server actions into actions module
